Keep direct policy preflight within architecture budget
